+++ src/core/astParse.ts
@@ -1,11 +1,9 @@
 import { parse } from "@babel/parser";
 import traverse from "@babel/traverse";
 import dealNewExpression from './ast/newExpression';
-// import dealVariableDeclarator from './ast/variableDeclarator';
 import dealCallExpression from './ast/callExpression';
 import handleTypeAnnotation from '../utils/tsTypeToAstNode';
 import { DiagnosticPayload } from "./diagnostic/diagnosticTypes";
-// import dealFucExpression from './ast/functionExpression';
 
 /**
  * 分析code
@@ -81,7 +79,7 @@ export function analyzeCode(code: string, url: string, startLine?: number) {
       }
     },
     // 处理所有new方法调用表达式
-    NewExpression: (path: CalleeType) => dealNewExpression(path, code, diagnosticsCallBack),
+    NewExpression: (path: CalleeType) => dealNewExpression(path, code, diagnosticsCallBack, startLine),
     // 分析方法调用表达式
     CallExpression: (path: CalleeType) => dealCallExpression(path, code, diagnosticsCallBack, variableTypes[url], startLine),
     OptionalCallExpression: (path: CalleeType) => dealCallExpression(path, code, diagnosticsCallBack, variableTypes[url], startLine),